refactor(ccj_postprocess): update 1.7.4, Refactor the repository
