+++ Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "ccj_postprocess"
-version = "1.7.3"
+version = "1.7.4"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-serde = "*"
+serde = { version = "*", features = ["derive"] }
 serde_derive = "*"
 serde_json = "*"
 clap = { version = "4.*", features = ["derive"] }
