+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod postprocess_config;
+pub mod arg_parser;
+pub mod compile_commands;
